update store page

+++ webforces/forms.py
@@ -7,7 +7,7 @@ class NewAlgForm(forms.Form):
     description = forms.CharField(label="",
                                   widget=forms.Textarea(attrs={'id': 'NewAlgDescription',
                                                                'placeholder': 'Description'}))
-    cost = forms.FloatField(min_value=0, label="",
-                            widget=forms.TextInput(attrs={'id': 'NewAlgCost', 'placeholder': 'Cost'}))
+    cost = forms.IntegerField(min_value=0, label="",
+                              widget=forms.TextInput(attrs={'id': 'NewAlgCost', 'placeholder': 'Cost'}))
     source = forms.CharField(label="",
                              widget=forms.Textarea(attrs={'id': 'NewAlgSource', 'placeholder': 'Source'}))
